Refactor code for robot movement functions
This commit refactors the code for the forward, backward, and turn_right functions in the main.py file. The functions now accept an additional parameter, previous_light_state, which is used for obstacle avoidance. The code for obstacle_avoidance function is also modified to improve the speech output when an obstacle is detected.